requirements.txt
Cython>=0.29.14
matplotlib>=3.1.2
numpy>=1.18.1
pandarallel>=1.4.5
pandas>=1.0.0
PyYAML>=5.2
scipy>=1.3.1
statsmodels>=0.11.0
